+++ includes/entidades/personagens/Personagem.hpp
@@ -16,7 +16,7 @@ namespace Entidades
             float vely;
 
         public:
-            Personagem(int posX = 0, int posY = 0, int tamanhoX = 0, int tamanhoY = 0, int n = 0);
+            Personagem(float posX = 0.0f, float posY = 0.0f, float tamanhoX = 0.0f, float tamanhoY = 0.0f, int n = 0);
             virtual ~Personagem();
 
             // void salvarDataBuffer();
